[Refactor] LIVD-439 - PR 리뷰 피드백 반영
List 네이밍 정리, Store 테스트 들여쓰기 통일, Sorter 파일 분리를 적용한다.
