+++ Projects/HomeFeature/Sources/Home/Model/CalendarDayScheduleItem.swift
@@ -54,33 +54,3 @@ extension CalendarDayScheduleItem {
         static let tbaLabel = "추후 발표"
     }
 }
-
-// MARK: - Sorter
-
-enum CalendarDayScheduleSorter {
-    static func sorted(_ items: [CalendarDayScheduleItem]) -> [CalendarDayScheduleItem] {
-        items.sorted { lhs, rhs in
-            let lhsRank = sortRank(lhs)
-            let rhsRank = sortRank(rhs)
-            if lhsRank != rhsRank {
-                return lhsRank < rhsRank
-            }
-
-            if lhsRank == 0, let lhsTime = lhs.time, let rhsTime = rhs.time, lhsTime != rhsTime {
-                return lhsTime < rhsTime
-            }
-
-            return lhs.title.localizedStandardCompare(rhs.title) == .orderedAscending
-        }
-    }
-
-    private static func sortRank(_ item: CalendarDayScheduleItem) -> Int {
-        if item.isCancelled {
-            return 2
-        }
-        if item.time == nil {
-            return 1
-        }
-        return 0
-    }
-}
